ignore gitkeep in skripte

+++ .Skripte/create_result.py
@@ -45,18 +45,19 @@ if __name__ == "__main__":
     CONFIG_PATH = config["PATHS"]["BASE"] + "/Configs"
 
     for file in os.listdir(CONFIG_PATH):
-        result = ""
-        tmp_config = f"{CONFIG_PATH}/{file}"
-        headers = extract_headings_from_markdown(tmp_config)
-        for h in headers:
-            counter = h.count("#")
-            file_name = h.split("# ")[1]
-            content = find_content_under_heading(config["PATHS"]["BASE"]+"/Zettel/"+file_name+".md", "Inhalt")
-            result = result + h + "\n"
-            if len(content) > 0:
-                for x in content:
-                    row_end = "\n" if not x.endswith("\n") else ""
-                    result = result + x + row_end
+        if file != ".gitkeep":
+            result = ""
+            tmp_config = f"{CONFIG_PATH}/{file}"
+            headers = extract_headings_from_markdown(tmp_config)
+            for h in headers:
+                counter = h.count("#")
+                file_name = h.split("# ")[1]
+                content = find_content_under_heading(config["PATHS"]["BASE"]+"/Zettel/"+file_name+".md", "Inhalt")
+                result = result + h + "\n"
+                if len(content) > 0:
+                    for x in content:
+                        row_end = "\n" if not x.endswith("\n") else ""
+                        result = result + x + row_end
 
-        with open(config["PATHS"]["BASE"]+"/Results/"+file, "w", encoding='utf-8') as file:
-            file.write(result)
+            with open(config["PATHS"]["BASE"]+"/Results/"+file, "w", encoding='utf-8') as file:
+                file.write(result)
